Refactors code to put ResultsList into separate component
